requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
python-dotenv
pydantic

# Dev
flake8
pytest
